feat(wavefront): full calculateDirectLightingUnified with inline shadows
Major quality improvement — use the monolithic's full direct lighting
function with inline shadow rays instead of simplified NEE:

- Request maxStorageBuffersPerShaderStage: 10 from WebGPU adapter
- Add bvhBuffer + triangleBuffer to Shade kernel (10 bindings total)
- Replace inline env IS NEE with calculateDirectLightingUnified()
  (handles ALL light types: directional, point, spot, area + env IS
  with full multi-strategy MIS and inline shadow traversal)
- Fix giScale: 1.0 at bounce 0, globalIlluminationIntensity for bounce>0
- Fix miss handler: match monolithic (no double MIS weighting)
- Fix readHitMaterialIndex: wrap .and() result in uint() to avoid
  WGSL bool type errors in clamp/arithmetic
- Shade first-dispatch: ~1450ms (one-time compilation cost)

Result: dramatically closer to monolithic quality. Camera body shows
proper contrast, shadows, metallic reflections. Still slightly brighter
due to uncached material classification in calculateIndirectLighting.

+++ src/core/Processor/PackedRayBuffer.js
@@ -11,7 +11,7 @@
  * separate TSL nodes that bind to the SAME GPU data with different access modes.
  */
 
-import { storage, uintBitsToFloat, floatBitsToUint, vec4 } from 'three/tsl';
+import { storage, uintBitsToFloat, floatBitsToUint, vec4, uint } from 'three/tsl';
 import { StorageInstancedBufferAttribute } from 'three/webgpu';
 
 // ─── Stride constants ───────────────────────────────────────────────────
@@ -260,7 +260,7 @@ export const readHitNormal = ( buf, id ) =>
 	buf.element( id.mul( HIT_STRIDE ).add( HIT.NORMAL_MAT ) ).xyz;
 
 export const readHitMaterialIndex = ( buf, id ) =>
-	floatBitsToUint( buf.element( id.mul( HIT_STRIDE ).add( HIT.NORMAL_MAT ) ).w ).and( 0xFFFF );
+	uint( floatBitsToUint( buf.element( id.mul( HIT_STRIDE ).add( HIT.NORMAL_MAT ) ).w ).and( 0xFFFF ) );
 
 export const readHitMeshIndex = ( buf, id ) =>
 	floatBitsToUint( buf.element( id.mul( HIT_STRIDE ).add( HIT.NORMAL_MAT ) ).w ).shiftRight( 16 );
